state_changer: moved class to it's own file for testing
